// Gerdau/web/src/redux/task/types.ts
export enum TaskActionTypes {
  FETCH_TASK = "FETCH_TASK",
  FETCH_TASK_SUCCESS = "FETCH_TASK_SUCCESS",
  FETCH_TASK_ERROR = "FETCH_TASK_ERROR",

  FETCH_TASKS = "FETCH_TASKS",
  FETCH_TASKS_SUCCESS = "FETCH_TASKS_SUCCESS",
  FETCH_TASKS_ERROR = "FETCH_TASKS_ERROR",

  FETCH_TASKS_BY_USER = "FETCH_TASKS_BY_USER",
  FETCH_TASKS_BY_USER_SUCCESS = "FETCH_TASKS_BY_USER_SUCCESS ",
  FETCH_TASKS_BY_USER_ERROR = "FETCH_TASKS_BY_USER_ERROR ",
}